Move labels and handles to presentational
